refactor: 폴더 구조 변경
- event는 model 폴더에, entity는 entity 폴더에 구분할 수 있도록 변경했습니다.
