+++ src/main/java/com/petitapetit/miml/domain/notification/entity/Notification.java
@@ -1,4 +1,4 @@
-package com.petitapetit.miml.domain.notification.model;
+package com.petitapetit.miml.domain.notification.entity;
 
 import com.petitapetit.miml.domain.notification.TempUser;
 import lombok.Getter;
