+++ src/test/java/com/petitapetit/miml/notification/NotificationServiceTest.java
@@ -5,15 +5,15 @@ import com.petitapetit.miml.domain.notification.TempArtist;
 import com.petitapetit.miml.domain.notification.TempSong;
 import com.petitapetit.miml.domain.notification.TempUser;
 import com.petitapetit.miml.domain.notification.TempUserRepository;
-import com.petitapetit.miml.domain.notification.model.FriendRequestedNotification;
-import com.petitapetit.miml.domain.notification.model.Notification;
-import com.petitapetit.miml.domain.notification.model.SharePlaylistRequestedNotification;
-import com.petitapetit.miml.domain.notification.model.SongAddedNotification;
+import com.petitapetit.miml.domain.notification.entity.FriendRequestedNotification;
+import com.petitapetit.miml.domain.notification.entity.Notification;
+import com.petitapetit.miml.domain.notification.entity.SharePlaylistRequestedNotification;
+import com.petitapetit.miml.domain.notification.entity.SongAddedNotification;
 import com.petitapetit.miml.domain.notification.repository.NotificationRepository;
-import com.petitapetit.miml.domain.notification.service.FriendRequestedEvent;
+import com.petitapetit.miml.domain.notification.model.FriendRequestedEvent;
 import com.petitapetit.miml.domain.notification.service.NotificationService;
-import com.petitapetit.miml.domain.notification.service.SharePlaylistRequestedEvent;
-import com.petitapetit.miml.domain.notification.service.SongAddedEvent;
+import com.petitapetit.miml.domain.notification.model.SharePlaylistRequestedEvent;
+import com.petitapetit.miml.domain.notification.model.SongAddedEvent;
 import com.petitapetit.miml.test.ServiceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
